# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ALPHA Control Plane API
  version: 0.1.0
  description: |
    The ALPHA control plane handles authentication, organization management,
    BYO-key storage, task tracking, usage metering, Stripe billing, and the
    skills marketplace.

    All endpoints (except /healthz and /v1/auth/*) require a Bearer JWT token
    in the Authorization header.

    Rate limits: 10 req/min for auth endpoints, 100 req/min for all others.
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.ALPHA.dev
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MarketplaceListing:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        description:
          type: string
        author:
          type: string
        review:
          type: string
          enum: [verified, reviewed, experimental]
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        install_count:
          type: integer
        rating:
          type: number
        permissions:
          type: array
          items:
            type: string
security:
  - BearerAuth: []
paths:
...
  /v1/marketplace:
    get:
      summary: Browse the curated marketplace
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: review
          in: query
          schema:
            type: string
            enum: [verified, reviewed, experimental]
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Marketplace listings
          content:
            application/json:
              schema:
                type: object
                properties:
                  listings:
                    type: array
                    items:
                      $ref: "#/components/schemas/MarketplaceListing"
                  total:
                    type: integer
  /v1/marketplace/{id}:
    get:
      summary: Get full listing details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Listing details
        "404":
          description: Not found
  /v1/marketplace/submit:
    post:
      summary: Submit a skill to the marketplace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manifest, category, tags, readme]
              properties:
                manifest:
                  type: object
                category:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
                readme:
                  type: string
                price_monthly:
                  type: number
                  default: 0
      responses:
        "201":
          description: Submission created
...